Rename to string_gf
