+++ arm9/arm9.lsf
@@ -79,7 +79,7 @@ Static arm9
     Object unk_02020C44.o
     Object unk_0202134C.o
     Object unk_02021590.o
-    Object unk_02021934.o
+    Object string_gf.o
     Object unk_02021FF8.o
     Object unk_02022244.o
     Object unk_02022294.o
