+++ arm9/src/font.c
@@ -6,7 +6,7 @@
 
 #include "gf_gfx_loader.h"
 #include "render_text.h"
-#include "string16.h"
+#include "string_gf.h"
 #include "string_control_code.h"
 #include "text.h"
 #include "unk_02021590.h"
