+++ arm9/src/string_gf.c
@@ -1,7 +1,7 @@
 #include "global.h"
 
 #include "heap.h"
-#include "string16.h"
+#include "string_gf.h"
 #include "string_util.h"
 
 #define ASSERT_STR16(_str) ({ GF_ASSERT(_str != NULL); GF_ASSERT(_str->magic == STR16_MAGIC); })
