+++ include/string_gf.h
