+++ include/unk_02064E4C.h
@@ -2,7 +2,7 @@
 #define POKEDIAMOND_UNK_02064E4C_H
 
 #include "msgdata.h"
-#include "string16.h"
+#include "string_gf.h"
 
 void sub_02064E4C(struct MsgData *msgData, u16 map_sec, struct String *dest);
 void sub_02064E60(u32 map_no, enum HeapID heapID, struct String *dest);
